remove Factory class
