+++ phpstan.src.neon
@@ -3,6 +3,6 @@ includes:
     - vendor/phpstan/phpstan-phpunit/rules.neon
 parameters:
     ignoreErrors:
-        - '#Strict comparison using === between false and string will always evaluate to false\.#'
         - '#Parameter \#1 \$components of static method League\\Uri\\Uri::createFromComponents\(\) expects array, array<string, int\|string\|null>\|string\|null given\.#'
-    reportUnmatchedIgnoredErrors: false
+        - '#Method League\\Uri\\Uri::create\(\) should return League\\Uri\\RFC3986UriInterface but returns League\\Uri\\RFC3986UriInterface\|Psr\\Http\\Message\\UriInterface.#'
+    reportUnmatchedIgnoredErrors: true
