+++ phpstan.tests.neon
@@ -8,9 +8,6 @@ parameters:
         - '#Parameter \#2 \$base_uri of static method League\\Uri\\Resolver::resolve\(\) expects League\\Uri\\RFC3986UriInterface\|Psr\\Http\\Message\\UriInterface, string given\.#'
         - '#Parameter \#1 \$uri of static method League\\Uri\\Resolver::relativize\(\) expects League\\Uri\\RFC3986UriInterface\|Psr\\Http\\Message\\UriInterface, string given\.#'
         - '#Parameter \#2 \$base_uri of static method League\\Uri\\Resolver::relativize\(\) expects League\\Uri\\RFC3986UriInterface\|Psr\\Http\\Message\\UriInterface, string given\.#'
-        - '#Parameter \#1 \$port of method League\\Uri\\Uri::withPort\(\) expects int\|null, string given\.#'
-        - '#Parameter \#1 \$path of method League\\Uri\\RFC3986UriInterface::withPath\(\) expects string, DateTime\|false given\.#'
         - '#Parameter \#[\d]+ \$[A-Za-z0-9]+ of static method PHPUnit\\Framework\\Assert::assertJsonStringEqualsJsonString\(\) expects string, string\|false given\.#'
         - '#Parameter \#1 \$fragment of method League\\Uri\\Http::withFragment\(\) expects string, array given\.#'
-        - '#Parameter \#1 \$path of method League\\Uri\\RFC3986UriInterface::withPath\(\) expects string, null given\.#'
-    reportUnmatchedIgnoredErrors: false
+    reportUnmatchedIgnoredErrors: true
